refactor(dashboard): extract the AI session stack section
Fourth composition-section extraction (B1): provider services, the provider
registry, read coordinator, alias/profile stores and controllers, and the
codex profile support probe move into src/dashboard/sections/aiSessionStack.ts.
No behavior change.
